project_requirements.txt
torch>=2.0.0
torchvision>=0.15.0
matplotlib>=3.5.0
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=9.0.0
flask>=2.0.0
flask-sqlalchemy>=3.0.0
psycopg2-binary>=2.9.0
gunicorn>=20.0.0
pyyaml>=6.0.0
pandas>=1.3.0
requests>=2.25.0
werkzeug>=2.0.0
email-validator>=2.0.0